batch sum-checks (#161)
